Keyboard input, improved controller equation

+++ main/Validators.cs
@@ -5,18 +5,18 @@ namespace CTT;
 
 public class RangeRule : ValidationRule
 {
-    public int Min { get; set; }
-    public int Max { get; set; }
+    public double Min { get; set; }
+    public double Max { get; set; }
 
     public override ValidationResult Validate(object value, CultureInfo cultureInfo)
     {
-        int valueNum = 0;
+        double valueNum = 0;
         string valueStr = (string)value;
 
         try
         {
             if (valueStr.Length > 0)
-                valueNum = int.Parse(valueStr);
+                valueNum = double.Parse(valueStr);
         }
         catch (Exception e)
         {
@@ -25,8 +25,7 @@ public class RangeRule : ValidationRule
 
         if ((valueNum < Min) || (valueNum > Max))
         {
-            return new ValidationResult(false,
-              $"Please enter a number in the range: {Min}-{Max}.");
+            return new ValidationResult(false, $"Please enter a number in the range: {Min}-{Max}.");
         }
         return ValidationResult.ValidResult;
     }
